ON-617 Content editorial feeds terms -- Update migration to check for new props before creating new index

+++ migrations/legacy/20190409120000-content-freq-editorial-feeds/migration.sh
@@ -27,37 +27,41 @@ printf "\n\n\n\n"
 newProps="\"featured\": {\"type\": \"boolean\"}, \"editorialFeeds\": {\"type\": \"object\", \"dynamic\": \"true\"}";
 
 index=$(curl -X GET "$es:9200/_cat/indices?pretty&s=index:desc" 2>/dev/null | sed -n 's/.*\(published-content[^ \n]*\).*/\1/p' | sed q);
-num=$(echo $index | sed 's/[^0-9]*//g');
-newIndex="$(echo $index | sed 's/[0-9]*//g')$((num+1))";
-alias=$(echo $index | sed 's/\(.*\)_.*/\1/g');
-mappings=$(curl -X GET "$es:9200/$index/_mappings" 2>/dev/null | sed -n "s/\"properties\":{/\"properties\":{$newProps,/p" | sed -n "s/{\"$index\":{\(.*\)}}/\1/p");
+mappings=$(curl -X GET "$es:9200/$index/_mappings");
 
-printf "\n\nCreating new index ($newIndex)...\n\n"
-curl -X GET "$es:9200/$index/_settings" > ./settings.json;
-node ./settings-update.js "$1";
-settings=$(cat ./settings.txt);
-curl -X PUT "$es:9200/$newIndex" -H 'Content-Type: application/json' -d "{$settings,$mappings}";
-rm ./settings.json;
-rm ./settings.txt;
+if [[ $mappings != *"featured"* && $mappings != *"editorialFeeds"* ]]; then
+  num=$(echo $index | sed 's/[^0-9]*//g');
+  newIndex="$(echo $index | sed 's/[0-9]*//g')$((num+1))";
+  alias=$(echo $index | sed 's/\(.*\)_.*/\1/g');
+  newMappings=$(curl -X GET "$es:9200/$index/_mappings" 2>/dev/null | sed -n "s/\"properties\":{/\"properties\":{$newProps,/p" | sed -n "s/{\"$index\":{\(.*\)}}/\1/p");
 
-printf "\r\n\r\nCopying old index data ($index) to new index ($newIndex)...\n\n"
-curl -X POST "$es:9200/_reindex" -H 'Content-Type: application/json' -d "
-{
-  \"source\": {
-    \"index\": \"$index\"
-  },
-  \"dest\": {
-    \"index\": \"$newIndex\"
-  }
-}";
+  printf "\n\nCreating new index ($newIndex)...\n\n"
+  curl -X GET "$es:9200/$index/_settings" > ./settings.json;
+  node ./settings-update.js "$1";
+  settings=$(cat ./settings.txt);
+  curl -X PUT "$es:9200/$newIndex" -H 'Content-Type: application/json' -d "{$settings,$newMappings}";
+  rm ./settings.json;
+  rm ./settings.txt;
 
-sleep 1;
+  printf "\r\n\r\nCopying old index data ($index) to new index ($newIndex)...\n\n"
+  curl -X POST "$es:9200/_reindex" -H 'Content-Type: application/json' -d "
+  {
+    \"source\": {
+      \"index\": \"$index\"
+    },
+    \"dest\": {
+      \"index\": \"$newIndex\"
+    }
+  }";
 
-printf "\n\nRemoving old alias and adding new ($alias)...\n\n"
-curl -X POST "$es:9200/_aliases" -H 'Content-Type: application/json' -d "
-{
-    \"actions\" : [
-        { \"remove\" : { \"index\" : \"$index\", \"alias\" : \"$alias\" } },
-        { \"add\" : { \"index\" : \"$newIndex\", \"alias\" : \"$alias\" } }
-    ]
-}";
+  sleep 1;
+
+  printf "\n\nRemoving old alias and adding new ($alias)...\n\n"
+  curl -X POST "$es:9200/_aliases" -H 'Content-Type: application/json' -d "
+  {
+      \"actions\" : [
+          { \"remove\" : { \"index\" : \"$index\", \"alias\" : \"$alias\" } },
+          { \"add\" : { \"index\" : \"$newIndex\", \"alias\" : \"$alias\" } }
+      ]
+  }";
+fi
